minor clean-up in `ManagedStrategy`
