+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/strategy/managed/ManagedStrategy.kt
@@ -632,11 +632,8 @@ internal abstract class ManagedStrategy(
         if (currentThreadId < nScenarioThreads) return
 
         onThreadStart(currentThreadId)
-        val methodId = context.createAndRegisterMethodDescriptor(
-            className = "java.lang.Thread",
-            methodName = "run",
-            methodType = Types.MethodType(Types.VOID_TYPE)
-        ).id
+
+        val methodId = context.getThreadRunMethodId()
         if (currentExecutionPart !== VALIDATION && !threadScheduler.isAborted(currentThreadId)) {
             loopDetector.onMethodEnter(
                 threadId = currentThreadId,
@@ -673,16 +670,13 @@ internal abstract class ManagedStrategy(
         // scenario threads are handled separately by the runner itself
         if (currentThreadId < nScenarioThreads) return
 
-        val methodId = context.getOrCreateMethodId(
-            className = "java.lang.Thread",
-            methodName = "run",
-            methodType = Types.MethodType(Types.VOID_TYPE)
-        )
         val eventId = getNextEventId()
         val threadRunTracePoint = callStackTrace[currentThreadId]?.firstOrNull()?.tracePoint
         val threadEndTracePoint = threadRunTracePoint?.let { MethodReturnTracePoint(context, eventId, it) }
         if (threadEndTracePoint != null) traceCollector?.addTracePoint(threadEndTracePoint)
         disableAnalysis()
+
+        val methodId = context.getThreadRunMethodId()
         if (currentExecutionPart !== VALIDATION && !threadScheduler.isAborted(currentThreadId)) {
             loopDetector.onMethodExit(
                 threadId = currentThreadId,
@@ -693,6 +687,7 @@ internal abstract class ManagedStrategy(
             )
         }
         loopDetector.resetThread(currentThreadId)
+
         onThreadFinish(currentThreadId)
     }
 
@@ -712,11 +707,7 @@ internal abstract class ManagedStrategy(
         // scenario threads are handled separately by the runner itself
         if (currentThreadId < nScenarioThreads) return
 
-        val methodId = context.getOrCreateMethodId(
-            className = "java.lang.Thread",
-            methodName = "run",
-            methodType = Types.MethodType(Types.VOID_TYPE)
-        )
+        val methodId = context.getThreadRunMethodId()
         if (currentExecutionPart !== VALIDATION && !threadScheduler.isAborted(currentThreadId)) {
             loopDetector.onMethodExit(
                 threadId = currentThreadId,
@@ -727,6 +718,7 @@ internal abstract class ManagedStrategy(
             )
         }
         loopDetector.resetThread(currentThreadId)
+
         // check if the exception is internal
         if (isLincheckInternalException(exception)) {
             onInternalException(currentThreadId, exception)
@@ -831,16 +823,16 @@ internal abstract class ManagedStrategy(
      * Aborts all threads in case if all `TestThread`s are in state `FINISHED`
      * and all running user threads are `LiveLocked` or `Parked`.
      *
-     * This method is expected to be called before [onSwitchPoint] method in
-     * order to abort execution and not to create unnecessary switch points.
+     * This method is expected to be called before [onSwitchPoint] method
+     * in order to abort execution and not to create unnecessary switch points.
      *
      * Since this method only can abort executions with `LiveLocked` or `Parked`
      * user-threads and `Finished` test threads, then invocations of this method are
-     * only meaningful near the corresponding strategy hooks
-     * ([ManagedStrategy.processLoopDetectorDecision], [ManagedStrategy.park],
-     * and [ManagedStrategy.onThreadFinish] where potentially test thread could finish).
+     * only meaningful near the corresponding strategy hooks,
+     * such as [ManagedStrategy.onLoopIteration], [ManagedStrategy.park], [ManagedStrategy.onThreadFinish], etc.;
+     * where potentially test thread could finish.
      *
-     * @param threadId id of thread that invoked this method.
+     * @param threadId id of the thread that invoked this method.
      * @param blockingReason blocking reason of invoking thread (determined by strategy) if exists.
      */
     private fun tryAbortingUserThreads(threadId: Int, blockingReason: BlockingReason?) {
@@ -957,6 +949,7 @@ internal abstract class ManagedStrategy(
             callType = MethodCallTracePoint.CallType.ACTOR,
         )
         traceCollector?.addTracePointInternal(tracePoint)
+
         if (currentExecutionPart !== VALIDATION && !threadScheduler.isAborted(iThread)) {
             loopDetector.onMethodEnter(
                 threadId = iThread,
@@ -966,6 +959,7 @@ internal abstract class ManagedStrategy(
                 params = actor.arguments.toTypedArray(),
             )
         }
+
         enableAnalysis()
     }
 
@@ -980,15 +974,10 @@ internal abstract class ManagedStrategy(
         val actorId = currentActorId[iThread]!!
         val scenario = runner.scenario
         val actor = if (actorId < scenario.threads[iThread].size) scenario.threads[iThread][actorId]
-        else runner.validationFunction
+                    else runner.validationFunction
         check(actor != null) { "Could not find current actor" }
-        val methodDescriptor = getAsmMethod(actor.method).descriptor
 
-        val methodId = context.getOrCreateMethodId(
-            className = actor.method.declaringClass.name.toCanonicalClassName(),
-            methodName = actor.method.name,
-            methodType = Types.convertAsmMethodType(methodDescriptor)
-        )
+        val methodId = context.getActorMethodId(actor)
         if (currentExecutionPart !== VALIDATION && !threadScheduler.isAborted(iThread)) {
             loopDetector.onMethodExit(
                 threadId = iThread,
@@ -998,6 +987,7 @@ internal abstract class ManagedStrategy(
                 result = null,
             )
         }
+
         val actorStartTracePoint = traceCollector?.trace
                 ?.filterIsInstance<MethodCallTracePoint>()
                 ?.firstOrNull { it.isActor && it.actorId == currentActorId[iThread] && it.iThread == iThread }
@@ -1337,7 +1327,6 @@ internal abstract class ManagedStrategy(
             memoryTracker!!.beforeRead(threadId, codeLocation, location)
             resultInterceptor?.interceptResult(memoryTracker!!.interceptReadResult(threadId))
         }
-
         return
     }
 
@@ -1363,7 +1352,6 @@ internal abstract class ManagedStrategy(
             memoryTracker!!.beforeRead(threadId, codeLocation, location)
             resultInterceptor?.interceptResult(memoryTracker!!.interceptReadResult(threadId))
         }
-
         return
     }
 
@@ -1883,7 +1871,7 @@ internal abstract class ManagedStrategy(
 
         val threadId = threadScheduler.getCurrentThreadId()
 
-// TODO: check what result to pass
+        // TODO: check what result to pass
         if (currentExecutionPart !== VALIDATION && !threadScheduler.isAborted(threadId)) {
             loopDetector.onMethodExit(
                 threadId = threadId,
@@ -1893,6 +1881,7 @@ internal abstract class ManagedStrategy(
                 result = null,
             )
         }
+
         // check if the called method is an atomics API method
         // (e.g., Atomic classes, AFU, VarHandle memory access API, etc.)
         val atomicMethodDescriptor = getAtomicMethodDescriptor(receiver, methodDescriptor.methodName)
@@ -1952,7 +1941,7 @@ internal abstract class ManagedStrategy(
         val threadId = threadScheduler.getCurrentThreadId()
 
 
-    // TODO: check what result to pass
+        // TODO: check what result to pass
         if (currentExecutionPart !== VALIDATION && !threadScheduler.isAborted(threadId)) {
             loopDetector.onMethodExit(
                 threadId = threadId,
@@ -1962,6 +1951,7 @@ internal abstract class ManagedStrategy(
                 result = null,
             )
         }
+
         // check if the called method is an atomics API method
         // (e.g., Atomic classes, AFU, VarHandle memory access API, etc.)
         val atomicMethodDescriptor = getAtomicMethodDescriptor(receiver, methodDescriptor.methodName)
@@ -2049,6 +2039,7 @@ internal abstract class ManagedStrategy(
                 traceCollector!!.addStateRepresentation()
             }
         }
+
         if (currentExecutionPart !== VALIDATION && !threadScheduler.isAborted(threadId)) {
             loopDetector.onMethodExit(
                 threadId = threadId,
@@ -2074,6 +2065,7 @@ internal abstract class ManagedStrategy(
                 traceCollector!!.addStateRepresentation()
             }
         }
+
         if (currentExecutionPart !== VALIDATION && !threadScheduler.isAborted(threadId)) {
             loopDetector.onMethodExit(
                 threadId = threadId,
