+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/util/Threads.kt
@@ -10,6 +10,10 @@
 
 package org.jetbrains.kotlinx.lincheck.util
 
+import org.jetbrains.lincheck.descriptors.Types
+import org.jetbrains.lincheck.trace.TraceContext
+import org.jetbrains.lincheck.trace.createAndRegisterMethodDescriptor
+
 typealias ThreadId = Int
 typealias ThreadMap<T> = Map<ThreadId, T>
 typealias MutableThreadMap<T> = MutableMap<ThreadId, T>
@@ -23,5 +27,12 @@ fun <T> threadMapOf(vararg pairs: Pair<ThreadId, T>): ThreadMap<T> =
 fun <T> mutableThreadMapOf(): MutableThreadMap<T> =
     mutableMapOf()
 
-fun <T> mutableThreadMapOf(vararg pairs: Pair<ThreadId, T>): MutableThreadMap<T> =
+fun <T> mutableThreadMapOf(vararg pairs: Pair<ThreadId, T>): ThreadMap<T> =
     mutableMapOf(*pairs)
+
+internal fun TraceContext.getThreadRunMethodId(): Int =
+    this.createAndRegisterMethodDescriptor(
+        className = "java.lang.Thread",
+        methodName = "run",
+        methodType = Types.MethodType(Types.VOID_TYPE)
+    ).id
